VideoDetail, UserDetail

+++ controllers/videoController.js
@@ -3,11 +3,11 @@ import Video from "../models/Video";
 
 export const home = async (req, res) => {
   try {
-    const videoList = await Video.find({}).sort({ _id: -1 });
-    res.render("home", { pageTitle: "Home", videoList });
+    const videos = await Video.find({}).sort({ _id: -1 });
+    res.render("home", { pageTitle: "Home", videos });
   } catch (error) {
     console.log(error);
-    res.render("home", { pageTitle: "Home", videoList: [] });
+    res.render("home", { pageTitle: "Home", videos: [] });
   }
 };
 
@@ -50,14 +50,13 @@ export const postUpload = async (req, res) => {
 };
 
 export const videoDetail = async (req, res) => {
-  console.log(req.user);
   const {
     params: { id }
   } = req;
 
   try {
     const video = await Video.findById(id).populate("creator");
-    console.log(video);
+
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     console.log(error);
@@ -69,10 +68,14 @@ export const getEditVideo = async (req, res) => {
   const {
     params: { id }
   } = req;
-
   try {
-    const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    const video = await Video.findById(id).populate("creator");
+
+    if (video.creator.id !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -96,7 +99,12 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id).populate("creator");
+    if (video.creator.id !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
